Modified map saving.

+++ src/orpg/editor/MapEditorWindow.java
@@ -50,7 +50,7 @@ public class MapEditorWindow extends JFrame implements Observer {
 	private BaseEditor baseEditor;
 
 	public MapEditorWindow(BaseEditor baseEditor) {
-		Map map = new Map(3, 3);
+		Map map = new Map((short) 3, (short) 3);
 
 		this.mapController = new MapController(map);
 		this.editorController = new MapEditorController(baseEditor,
@@ -85,8 +85,7 @@ public class MapEditorWindow extends JFrame implements Observer {
 		mapScrollPane.getVerticalScrollBar().setUnitIncrement(16);
 		mapScrollPane.setBackground(Color.black);
 		mapContainer.add(mapScrollPane);
-		mapContainer
-				.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
+		mapContainer.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
 
 		contentPane.add(mapContainer);
 		this.add(contentPane);
@@ -120,9 +119,9 @@ public class MapEditorWindow extends JFrame implements Observer {
 		// Build the tiles view
 		TilesView tilesView = null;
 		try {
-			tilesView = new TilesView(editorController,
-					ImageIO.read(new File(Constants.CLIENT_DATA_PATH
-							+ "gfx/tiles.png").toURI().toURL()));
+			tilesView = new TilesView(editorController, ImageIO.read(new File(
+					Constants.CLIENT_DATA_PATH + "gfx/tiles.png").toURI()
+					.toURL()));
 		} catch (MalformedURLException e) {
 			e.printStackTrace();
 			System.exit(1);
@@ -151,9 +150,8 @@ public class MapEditorWindow extends JFrame implements Observer {
 			public void itemStateChanged(ItemEvent e) {
 				JRadioButton button = (JRadioButton) e.getSource();
 				if (button.isSelected()) {
-					editorController
-							.setCurrentLayer(MapLayer.values()[Integer
-									.parseInt(button.getActionCommand())]);
+					editorController.setCurrentLayer(MapLayer.values()[Integer
+							.parseInt(button.getActionCommand())]);
 				}
 			}
 		};
@@ -186,8 +184,7 @@ public class MapEditorWindow extends JFrame implements Observer {
 		fileMenu.setMnemonic(KeyEvent.VK_F);
 		menuBar.add(fileMenu);
 
-		JMenuItem saveItem = new JMenuItem(
-				editorController.getSaveAction());
+		JMenuItem saveItem = new JMenuItem(editorController.getSaveAction());
 		saveItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_S,
 				KeyEvent.CTRL_DOWN_MASK));
 		fileMenu.add(saveItem);
@@ -197,14 +194,12 @@ public class MapEditorWindow extends JFrame implements Observer {
 		editMenu.setMnemonic(KeyEvent.VK_E);
 		menuBar.add(editMenu);
 
-		JMenuItem undoItem = new JMenuItem(
-				editorController.getUndoAction());
+		JMenuItem undoItem = new JMenuItem(editorController.getUndoAction());
 		undoItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_Z,
 				KeyEvent.CTRL_DOWN_MASK));
 		editMenu.add(undoItem);
 
-		JMenuItem redoItem = new JMenuItem(
-				editorController.getRedoAction());
+		JMenuItem redoItem = new JMenuItem(editorController.getRedoAction());
 		redoItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_Y,
 				KeyEvent.CTRL_DOWN_MASK));
 		editMenu.add(redoItem);
@@ -214,17 +209,16 @@ public class MapEditorWindow extends JFrame implements Observer {
 		viewMenu.setMnemonic(KeyEvent.VK_V);
 		menuBar.add(viewMenu);
 
-		JMenuItem zoomInItem = new JMenuItem(
-				editorController.getZoomInAction());
-		zoomInItem.setAccelerator(KeyStroke.getKeyStroke(
-				KeyEvent.VK_EQUALS, KeyEvent.CTRL_DOWN_MASK));
+		JMenuItem zoomInItem = new JMenuItem(editorController.getZoomInAction());
+		zoomInItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_EQUALS,
+				KeyEvent.CTRL_DOWN_MASK));
 
 		viewMenu.add(zoomInItem);
 
 		JMenuItem zoomOutItem = new JMenuItem(
 				editorController.getZoomOutAction());
-		zoomOutItem.setAccelerator(KeyStroke.getKeyStroke(
-				KeyEvent.VK_MINUS, KeyEvent.CTRL_DOWN_MASK));
+		zoomOutItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_MINUS,
+				KeyEvent.CTRL_DOWN_MASK));
 		viewMenu.add(zoomOutItem);
 
 		gridToggleMenuItem = new JCheckBoxMenuItem("Grid",
@@ -233,13 +227,12 @@ public class MapEditorWindow extends JFrame implements Observer {
 		gridToggleMenuItem.addItemListener(new ItemListener() {
 			@Override
 			public void itemStateChanged(ItemEvent e) {
-				editorController.setGridEnabled(gridToggleMenuItem
-						.getState());
+				editorController.setGridEnabled(gridToggleMenuItem.getState());
 			}
 		});
 
-		hoverPreviewToggleMenuItem = new JCheckBoxMenuItem(
-				"Mouse Preview", editorController.isHoverPreviewEnabled());
+		hoverPreviewToggleMenuItem = new JCheckBoxMenuItem("Mouse Preview",
+				editorController.isHoverPreviewEnabled());
 		viewMenu.add(hoverPreviewToggleMenuItem);
 		hoverPreviewToggleMenuItem.addItemListener(new ItemListener() {
 			@Override
